Moved CurrentMapLayer to GI
