+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Game/CarlaGameInstance.h
@@ -13,6 +13,7 @@
 #include "Carla/Server/CarlaServer.h"
 
 #include <compiler/disable-ue4-macros.h>
+#include "carla/rpc/MapLayer.h"
 #include <carla/rpc/OpendriveGenerationParameters.h>
 #include <compiler/enable-ue4-macros.h>
 
@@ -79,15 +80,27 @@ public:
   }
 
   void SetOpendriveGenerationParameters(
-      const carla::rpc::OpendriveGenerationParameters & Parameters) 
+      const carla::rpc::OpendriveGenerationParameters & Parameters)
   {
     GenerationParameters = Parameters;
   }
 
   const carla::rpc::OpendriveGenerationParameters&
-      GetOpendriveGenerationParameters() const 
+      GetOpendriveGenerationParameters() const
   {
     return GenerationParameters;
+  }
+
+  UFUNCTION(Category = "Carla Game Instance", BlueprintCallable)
+  void SetMapLayer(int32 MapLayer)
+  {
+    CurrentMapLayer = MapLayer;
+  }
+
+  UFUNCTION(Category = "Carla Game Instance", BlueprintCallable)
+  int32 GetCurrentMapLayer() const
+  {
+    return CurrentMapLayer;
   }
 
 private:
@@ -101,5 +114,8 @@ private:
   ACarlaRecorder *Recorder = nullptr;
 
   carla::rpc::OpendriveGenerationParameters GenerationParameters;
+
+  UPROPERTY(Category = "CARLA Game Instance", EditAnywhere)
+  int32 CurrentMapLayer = static_cast<int32>(carla::rpc::MapLayer::All);
 
 };
